+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 - feat: add aml provider coin kyt
+- feat: block hot wallet withdrawals while processing wallet withdrawals are pending (per user, per blockchain)
 
 ## [0.9.21] - 2026-07-24
 - feat: add aml scoring for deposit